# contracts/openapi/core-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Katha core-api",
    "description": "Public mobile + web API. Money is an append-only ledger; the client computes no prices or entitlements.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/home": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Home",
        "description": "The home rails. Anonymous callers get catalog order; a bearer gets\nevents-ranked Trending plus a \"Because you watched\" rail scored off the\nviewer's own progress (see app/recs.py).",
        "operationId": "home_v1_home_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "hi",
              "title": "Lang"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/series/{slug}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Series Detail",
        "operationId": "series_detail_v1_series__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Search",
        "description": "Catalog search (no auth). Series match on title / native title / tropes /\ngenres, case-insensitive substring; people match on the billed cast name.\nOnly served (live) series are searchable or listed under a person \u2014 the\nsame gate as the catalog listing, so an archived title never leaks.",
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CastMember": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "name",
          "role"
        ],
        "title": "CastMember"
      },
...
      "Episode": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_free": {
            "type": "boolean",
            "title": "Is Free"
          },
          "coin_price": {
            "type": "integer",
            "title": "Coin Price"
          }
        },
        "type": "object",
        "required": [
          "number",
          "title",
          "is_free",
          "coin_price"
        ],
        "title": "Episode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeResponse": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/HomeRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "HomeResponse"
      },
      "HomeRow": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesSummary"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "title",
          "series"
        ],
        "title": "HomeRow"
      },
...
      "SearchPerson": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesSummary"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "series"
        ],
        "title": "SearchPerson"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesSummary"
            },
            "type": "array",
            "title": "Series"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/SearchPerson"
            },
            "type": "array",
            "title": "People"
          }
        },
        "type": "object",
        "required": [
          "query",
          "series",
          "people"
        ],
        "title": "SearchResponse"
      },
      "SeriesDetail": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "episode_count": {
            "type": "integer",
            "title": "Episode Count"
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language"
          },
          "content_rating": {
            "type": "string",
            "title": "Content Rating",
            "default": "U/A 16+"
          },
          "cover_url": {
            "type": "string",
            "title": "Cover Url",
            "default": ""
          },
          "cover_wide_url": {
            "type": "string",
            "title": "Cover Wide Url",
            "default": ""
          },
          "synopsis": {
            "type": "string",
            "title": "Synopsis"
          },
          "title_native": {
            "type": "string",
            "title": "Title Native",
            "default": ""
          },
          "tropes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tropes"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastMember"
            },
            "type": "array",
            "title": "Cast"
          },
          "free_episode_count": {
            "type": "integer",
            "title": "Free Episode Count"
          },
          "episode_coin_price": {
            "type": "integer",
            "title": "Episode Coin Price"
          },
          "bundle_discount_pct": {
            "type": "integer",
            "title": "Bundle Discount Pct"
          },
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/Episode"
            },
            "type": "array",
            "title": "Episodes"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "genres",
          "episode_count",
          "primary_language",
          "synopsis",
          "free_episode_count",
          "episode_coin_price",
          "bundle_discount_pct",
          "episodes"
        ],
        "title": "SeriesDetail"
      },
      "SeriesSummary": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "episode_count": {
            "type": "integer",
            "title": "Episode Count"
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language"
          },
          "content_rating": {
            "type": "string",
            "title": "Content Rating",
            "default": "U/A 16+"
          },
          "cover_url": {
            "type": "string",
            "title": "Cover Url",
            "default": ""
          },
          "cover_wide_url": {
            "type": "string",
            "title": "Cover Wide Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "genres",
          "episode_count",
          "primary_language"
        ],
        "title": "SeriesSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}